Bugfix/linux support (#3)
* Fix baseUri when not webapplication

When the project using the library does not have a http context the baseUri of `/` would cause timeouts during navigation. By setting the value as `null` the PDF is generated as expected.

* Fix version mismatch

For Windows an update had been released for version 11.4.5. This works fine for Windows but when a project is being deployed to Linux it gave a runtime exception that version 11.4.5 could not be found while Linux is stated as version 11.4.1.

This was caused by NuGet restore using the highest version to apply for both. By declaring a range for Windows it allows NuGet to drop in version number to find a match for Linux that is available, thus version 11.4.1 is used.